FEAT: diffuse 기능 일부 추가
