+++ src/img/print_img.c
@@ -6,34 +6,11 @@
 /*   By: jhwang2 <jhwang2@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/07 17:32:47 by ahkiler           #+#    #+#             */
-/*   Updated: 2023/07/04 15:16:52 by jhwang2          ###   ########.fr       */
+/*   Updated: 2023/07/04 15:18:55 by jhwang2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minirt.h"
-
-double	max(double a, double b)
-{
-	if (a >= b)
-		return (a);
-	return (b);
-}
-
-u_int32_t	diffuse(t_data *data, t_rec *rec)
-{
-	double	diff;
-	double	r;
-	double	g;
-	double	b;
-
-	diff = max (v_dot (rec->n_vector,
-				v_unit (v_sub_vec (data->lights[0]->point,
-						rec->frag_point))), 0.0);
-	b = ((rec->color & 255));
-	g = (((rec->color >> 8) & 255) % 256);
-	r = (((rec->color >> 16) & 255) % 256);
-	return (0 | (int)(r * diff) << 16 | (int)(g * diff) << 8 | (int)(b * diff));
-}
 
 void	print_img(t_data *data)
 {
