+++ src/include/ray.h
@@ -6,7 +6,7 @@
 /*   By: jhwang2 <jhwang2@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/14 17:37:53 by jeelee            #+#    #+#             */
-/*   Updated: 2023/07/03 17:11:44 by jhwang2          ###   ########.fr       */
+/*   Updated: 2023/07/04 15:18:45 by jhwang2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,8 @@
 # include "./types.h"
 
 u_int32_t	create_trgb(t_data *data, t_rec *rec, int flag);
-
+u_int32_t	diffuse(t_data *data, t_rec *rec);
+double		max(double a, double b);
 int			hit_sphere(t_ray *ray, t_object *obj, double value[]);
 int			hit_plane(t_ray *ray, t_object *obj, double value[]);
 int			hit_cylinder(t_ray *ray, t_object *obj, double value[]);
